Make building of OpenCV optional (#77)
* Building OpenCV is now optional. Use option IBIS_INSTALL_OPENCV in superbuild to build OpenCV and IBIS_USE_OPENCV in the Ibis project if you want it to actually use OpenCV, which is required only from certain non-default plugins.

* ibismath.* opencv related functions and vtkopencvbridge.* to a new lib called IbisOpenCV.

* Moving to OpenCV 3.2 and now require exact version of the lib.
