+++ CMakeCommon/DependencyVersions.cmake
@@ -11,7 +11,7 @@ set( IBIS_ITK_SHORT_VERSION ${IBIS_ITK_MAJOR_VERSION}.${IBIS_ITK_MINOR_VERSION} 
 set( IBIS_ITK_LONG_VERSION ${IBIS_ITK_MAJOR_VERSION}.${IBIS_ITK_MINOR_VERSION}.${IBIS_ITK_BUILD_VERSION} )
 
 set( IBIS_OPENCV_MAJOR_VERSION 3 )
-set( IBIS_OPENCV_MINOR_VERSION 1 )
+set( IBIS_OPENCV_MINOR_VERSION 2 )
 set( IBIS_OPENCV_BUILD_VERSION 0 )
 set( IBIS_OPENCV_SHORT_VERSION ${IBIS_OPENCV_MAJOR_VERSION}.${IBIS_OPENCV_MINOR_VERSION} )
 set( IBIS_OPENCV_LONG_VERSION ${IBIS_OPENCV_MAJOR_VERSION}.${IBIS_OPENCV_MINOR_VERSION}.${IBIS_OPENCV_BUILD_VERSION} )
