+++ IbisLib/ibismath.h
@@ -13,15 +13,11 @@ See Copyright.txt or http://ibisneuronav.org/Copyright.html for details.
 
 #include "SVL.h"
 #include <vector>
-#include "opencv2/opencv.hpp"
 
 class vtkMatrix4x4;
 
-class IbisMath
+namespace IbisMath
 {
-
-public:
-
     // Scalar operations
     static double Average( std::vector<double> & samples );
     static double StdDev( std::vector<double> & samples, double average );
@@ -31,21 +27,10 @@ public:
     static Vec3 AverageVec3( std::vector<Vec3> & allVecs );
     static Vec3 StdDevVec3( std::vector<Vec3> & allVecs );
 
-    // cv::point operations
-    static cv::Point3f AverageCvPoint3f( std::vector<cv::Point3f> & all );
-    static cv::Point2f AverageCvPoint2f( std::vector<cv::Point2f> & all );
-
     // VTK mat 4 and SVL Vec3
     static Vec3 Translation( vtkMatrix4x4 * mat );
     static void MultMat4Vec3( vtkMatrix4x4 * mat, const Vec3 & in, Vec3 & out );
     static Vec3 MultMat4Point3( vtkMatrix4x4 * mat, const Vec3 & in );
-
-    // VTK mat 4 and Cv::Point3f
-    static cv::Point3f MultMat4CvPoint3f( vtkMatrix4x4 * mat, cv::Point3f & in );
-
-    // SVL vec and cv vec
-    static cv::Point3f Vec3ToCVPoint3f( Vec3 & v );
-    static Vec2 CVPoint2fToVec2( cv::Point2f & p );
 
     // VTK mat4 utilities
     static vtkMatrix4x4 * DuplicateMat4( vtkMatrix4x4 * in );
