+++ IbisPlugins/FiberNavigator/CMakeLists.txt
@@ -8,3 +8,6 @@ add_definitions( -DHAVE_ZLIB )
 
 # Create plugin
 DefinePlugin( "${PluginSrc}" "${PluginHdr}" "${PluginHdrMoc}" "${PluginUi}" )
+
+#add library specific to this plugin
+target_link_libraries( ${PluginName} z )
