+++ IbisSuperBuild/ibisExternalDependencies/CMakeLists.txt
@@ -36,19 +36,23 @@ ExternalProject_Add( ${itk_name}
   )
 
 # Get and build OpenCV
-set( opencv_name opencv-${IBIS_OPENCV_LONG_VERSION} )
-ExternalProject_Add( ${opencv_name} 
-    SOURCE_DIR ${external_project_dir}/${opencv_name}/src
-    GIT_REPOSITORY "https://github.com/opencv/opencv.git"
-    GIT_TAG ef0811018229c584c21dafa1bf9555c9f56a1d4e
-    BINARY_DIR ${external_project_dir}/${opencv_name}/build
-    BUILD_COMMAND   make
-    INSTALL_COMMAND make install 
-    CMAKE_ARGS -DCMAKE_INSTALL_PREFIX:PATH=${external_project_dir}/${opencv_name}/install -DCMAKE_OSX_SYSROOT:PATH=${CMAKE_OSX_SYSROOT} -DCMAKE_OSX_DEPLOYMENT_TARGET:STRING=${CMAKE_OSX_DEPLOYMENT_TARGET} -DCMAKE_BUILD_TYPE:STRING=${CMAKE_BUILD_TYPE} -DBUILD_PERF_TESTS:BOOL=FALSE -DBUILD_SHARED_LIBS:BOOL=FALSE -DBUILD_TESTS:BOOL=FALSE -DWITH_TIFF:BOOL=FALSE -DWITH_JASPER:BOOL=FALSE -DBUILD_DOCS:BOOL=FALSE -DBUILD_FAT_JAVA_LIB:BOOL=FALSE -DWITH_CUDA:BOOL=FALSE
-  )
+if( IBIS_INSTALL_OPENCV )
+    set( opencv_name opencv-${IBIS_OPENCV_LONG_VERSION} )
+    ExternalProject_Add( ${opencv_name}
+        SOURCE_DIR ${external_project_dir}/${opencv_name}/src
+        GIT_REPOSITORY "https://github.com/opencv/opencv.git"
+        GIT_TAG 70bbf17b133496bd7d54d034b0f94bd869e0e810
+        BINARY_DIR ${external_project_dir}/${opencv_name}/build
+        BUILD_COMMAND   make
+        INSTALL_COMMAND make install
+        CMAKE_ARGS -DCMAKE_INSTALL_PREFIX:PATH=${external_project_dir}/${opencv_name}/install -DCMAKE_OSX_SYSROOT:PATH=${CMAKE_OSX_SYSROOT} -DCMAKE_OSX_DEPLOYMENT_TARGET:STRING=${CMAKE_OSX_DEPLOYMENT_TARGET} -DCMAKE_BUILD_TYPE:STRING=${CMAKE_BUILD_TYPE} -DBUILD_PERF_TESTS:BOOL=FALSE -DBUILD_SHARED_LIBS:BOOL=FALSE -DBUILD_TESTS:BOOL=FALSE -DWITH_TIFF:BOOL=FALSE -DWITH_JASPER:BOOL=FALSE -DBUILD_DOCS:BOOL=FALSE -DBUILD_FAT_JAVA_LIB:BOOL=FALSE -DWITH_CUDA:BOOL=FALSE
+      )
+endif()
 
 add_custom_target( IbisExternalDependencies ALL )
 add_dependencies( IbisExternalDependencies ${vtk_name} )
 add_dependencies( IbisExternalDependencies ${itk_name} )
+if( IBIS_INSTALL_OPENCV )
 add_dependencies( IbisExternalDependencies ${opencv_name} )
+endif()
